requirements_pi.txt
# Raspberry Pi specific requirements for rover project
# Install with: pip3 install -r requirements_pi.txt

# Core web framework
Flask==2.3.3
Werkzeug==2.3.7

# Camera and image processing
opencv-python==4.8.1.78
numpy==1.24.3
picamera2==0.3.12

# GPIO control for Raspberry Pi
RPi.GPIO==0.7.1

# Development and testing
gpiozero==2.0

# Optional: For better performance
Pillow==10.0.1
